time in function, time ticks

+++ laba3/algo.h
@@ -9,10 +9,10 @@ using namespace std;
 int AnalizMaxStep(vector<tPoint> &vec, int step = 1);
 
 int SomeAlgo(vector<tPoint> &vec, const QPoint &start, const QPoint &end);
-int CdaAlgo(vector<tPoint> &vec, const QPoint &start, const QPoint &end);
-int BresenhamDoubleAlgo(vector<tPoint> &vec, const QPoint &start, const QPoint &end);
-int BresenhamIntAlgo(vector<tPoint> &vec, const QPoint &start, const QPoint &end);
-int BresenhamSmoothAlgo(vector<tPoint> &vec, const QPoint &start, const QPoint &end);
-int WuAlgo(vector<tPoint> &vec, const QPoint &start, const QPoint &end);
+double CdaAlgo(vector<tPoint> &vec, const QPoint &start, const QPoint &end);
+double BresenhamDoubleAlgo(vector<tPoint> &vec, const QPoint &start, const QPoint &end);
+double BresenhamIntAlgo(vector<tPoint> &vec, const QPoint &start, const QPoint &end);
+double BresenhamSmoothAlgo(vector<tPoint> &vec, const QPoint &start, const QPoint &end);
+double WuAlgo(vector<tPoint> &vec, const QPoint &start, const QPoint &end);
 
 #endif // ALGO_H
